udpated chrindex

+++ ft_chrdindex.c
@@ -6,13 +6,13 @@
 /*   By: pholster <pholster@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/02/03 22:38:37 by pholster       #+#    #+#                */
-/*   Updated: 2019/04/11 21:41:16 by pholster      ########   odam.nl         */
+/*   Updated: 2019/05/02 19:13:48 by pholster      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "includes/libft.h"
 
-int		ft_chrdindex(const char *str, const char *dlm)
+size_t		ft_chrdindex(const char *str, const char *dlm)
 {
 	size_t		i;
 	long long	bytes[2];
