+++ ft_chrindex.c
@@ -6,13 +6,13 @@
 /*   By: pholster <pholster@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/02/03 22:38:37 by pholster       #+#    #+#                */
-/*   Updated: 2019/05/01 15:51:57 by pholster      ########   odam.nl         */
+/*   Updated: 2019/05/02 19:13:52 by pholster      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "includes/libft.h"
 
-int		ft_chrindex(const char *str, int c)
+size_t		ft_chrindex(const char *str, int c)
 {
 	char	*ptr;
 
@@ -21,5 +21,5 @@ int		ft_chrindex(const char *str, int c)
 	ptr = ft_memchr(str, c, ft_strlen(str));
 	if (ptr == NULL)
 		return (-1);
-	return (ptr - str);
+	return ((size_t)(((unsigned long)ptr) - ((unsigned long)str)));
 }
